Module_3_Retrieval/Test/requirements.txt
torch
matplotlib
torchvision
neptune-client
tqdm
pandas
cudatoolkit
numpy


